requirements.txt
aiohttp==3.8.3
discord.py==2.1.0
dateparser==1.1.6
mystbin.py==5.1.0 # this caused errors for me before, if it does for you, just remove the ==5.1.0 part
better-ipc==2.0.1
humanize==4.4.0
psutil==5.9.4
requests==2.28.2
expr.py==0.3.0
playwright==1.29.1
jishaku==2.5.1
beautifulsoup4==4.11.1
python-Levenshtein==0.20.9
fuzzywuzzy==0.18.0
humanfriendly==10.0
sphobjinv==2.3.1
lxml==4.9.2
python-magic==0.4.27
nudenet==2.0.9
yt-dlp==2023.1.6
imagetext_py==2.1.2
Pillow==9.4.0aiohttp==3.8.3
discord.py==2.1.0
dateparser==1.1.6
mystbin.py==5.1.0 # this caused errors for me before, if it does for you, just remove the ==5.1.0 part
better-ipc==2.0.1
humanize==4.4.0
psutil==5.9.4
requests==2.28.2
expr.py==0.3.0
playwright==1.29.1
jishaku==2.5.1
beautifulsoup4==4.11.1
python-Levenshtein==0.20.9
fuzzywuzzy==0.18.0
humanfriendly==10.0
sphobjinv==2.3.1
lxml==4.9.2
python-magic==0.4.27
nudenet==2.0.9
yt-dlp==2023.1.6
imagetext_py==2.1.2
Pillow==9.4.0
expiringdict==1.2.2